Testing iteration weight being applied to simc

+++ src/simc_ana/iter_weight.py
@@ -3,7 +3,7 @@
 #
 # Description: Adapted from fortran code wt28_3.f
 # ================================================================
-# Time-stamp: "2023-09-18 15:33:33 trottar"
+# Time-stamp: "2023-09-18 15:34:03 trottar"
 # ================================================================
 #
 # Author:  Richard L. Trotta III <trotta@cua.edu>
@@ -93,7 +93,7 @@ def iter_weight(param_file, simc_root, inpDict, phi_setting):
     Weight_SIMC  = TBRANCH_SIMC.GetBranch("Weight")
 
     # Associate a variable with the branch
-    iweight = ROOT.Float(0)
+    iweight = ROOT.Double(0)
 
     # Create a new TBranch with the same name 'Weight' in a new TTree
     new_tree = ROOT.TTree("TBRANCH_SIMC", "Modified TTree with Weight")
